Forgot to (not to) filter mouse move events
On Windows, when I subclassed PyMouseEvent and instantiated that subclass passing "capture=True" to the constructor, my mouse would get stuck, as if I made "capture_move=True". That was because the MouseAll hook function would filter both the mouse click and move events only by checking the "self.capture" attribute at line 133, making the the mouse move events captured regardless of the value of self.capture_move. In order to correct that, I added line 115 to filter these events in the way it was intended